+++ integrations/awss3/v3/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `precision-soft/melody/integrations/awss3` will be docume
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v3.0.0] - 2026-06-15 - Initial Release — S3-Compatible Object Storage
+## [v3.0.0] - 2026-06-16 - Initial Release — S3-Compatible Object Storage
 
 ### Added